example/requirements.txt
certifi==2022.12.7
joblib==1.2.0
lightgbm==3.3.5
numpy==1.24.2
paho-mqtt==1.6.1
pandas==1.5.3
pickle4==0.0.1
pip==23.0.1
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.2
scipy==1.10.1
setuptools==65.6.3
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
wheel==0.38.4
